--- pages/test/main.module.sass
@@ -1,16 +1,0 @@
-.heatmap-page
-    margin: 2em 20%
-
-#map
-    height: 100%
-    width: 100%
-    aspect-ratio: 3/2
-    cursor: pointer
-    z-index: 100
-
-.map-area-container
-  width: 100% !important
-  height: 50vh !important
-
-.loading
-    padding-top: 100px
